requirements.txt
Flask~=2.0.2
flask-bootstrap
Flask-SQLAlchemy
WTForms~=2.1
gunicorn
python-dotenv